refactor: timer상수 파일 분리
