+++ frontend/src/constant/timer.ts
@@ -1,0 +1,4 @@
+export const SECOND = 1000;
+export const MINUTE = 60 * SECOND;
+export const HOUR = 60 * MINUTE;
+export const MAX_TIME = 4 * HOUR;
