+++ frontend/src/types/timer.ts
@@ -1,8 +1,3 @@
 export type ExpirationTime = {
     maxAge: string;
 };
-
-export const SECOND = 1000;
-export const MINUTE = 60 * SECOND;
-export const HOUR = 60 * MINUTE;
-export const MAX_TIME = 4 * HOUR;
